requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.4
psycopg2-binary==2.9.9
Flask-Bcrypt==1.0.1
PyJWT==2.8.0
Flask-CORS==4.0.0
marshmallow==3.21.1
marshmallow-sqlalchemy==0.29.0
python-dotenv==1.0.1
